feat(Wallet): Use `postMessage` for communication with extension wall… (#815)
* feat(Wallet): Use `postMessage` for communication with extension wallet through `Content Script`(bridge between page and extension)
Adjust example extension

* feat(Wallet): Remove `*` from postMessage,m use the same origin for extension and origin from connection for iframe wallet

* feat(Wallet): Wrap messages between `page` and `content script` in `{ type, data }` where is type is direction `to_aepp, to_waellet`. This was done because the message contain the same `origin`

* feat(Wallet): More generic direction handling. Fix LGTM

* feat(Wallet): Add debug mode for connections. Implement `ContentScriptBridge` stamp for content script. Adjust example apps

* refactor(Chain): handle time until tx is not added to mempool

* docs(connections): Add docs for connections stamps

* docs(connections): Add docs for content script bridge

* docs(AEPP<->Wallet): Add more docs

* docs(AEPP<->Wallet): Add more docs

* docs(AEPP<->Wallet): And more docs

* docs(AEPP<->Wallet): Finish API docs

* docs(AEPP<->Wallet): Fix aeppRpc

* feat(AEPP<->Wallet): Add validation for events handlers
Fix example aepp

* feat(AEPP<->Wallet): Fix extension example

* docs(AEPP<->Wallet): Draft for guide

* docs(AEPP<->Wallet): Add more text

* docs(AEPP<->Wallet): Add AEPP building guide using new integration API

* docs(AEPP<->Wallet): Regenerate lock files for example apps/extension
