+++ docs/api.md
@@ -26,6 +26,14 @@
   * [@aeternity/aepp-sdk/es/tx/builder/schema](api/tx/builder/schema.md)
   * [@aeternity/aepp-sdk/es/tx](api/tx.md)
   * [@aeternity/aepp-sdk/es/tx/tx](api/tx/tx.md)
+  * [@aeternity/aepp-sdk/es/utils/aepp-wallet-communication/connection/browser-runtime](api/utils/aepp-wallet-communication/connection/browser-runtime.md)
+  * [@aeternity/aepp-sdk/es/utils/aepp-wallet-communication/connection/browser-window-message](api/utils/aepp-wallet-communication/connection/browser-window-message.md)
+  * [@aeternity/aepp-sdk/es/utils/aepp-wallet-communication/connection](api/utils/aepp-wallet-communication/connection.md)
+  * [@aeternity/aepp-sdk/es/utils/aepp-wallet-communication/content-script-bridge](api/utils/aepp-wallet-communication/content-script-bridge.md)
+  * [@aeternity/aepp-sdk/es/utils/aepp-wallet-communication/rpc/aepp-rpc](api/utils/aepp-wallet-communication/rpc/aepp-rpc.md)
+  * [@aeternity/aepp-sdk/es/utils/aepp-wallet-communication/rpc/rpc-client](api/utils/aepp-wallet-communication/rpc/rpc-client.md)
+  * [@aeternity/aepp-sdk/es/utils/aepp-wallet-communication/rpc/wallet-rpc](api/utils/aepp-wallet-communication/rpc/wallet-rpc.md)
+  * [@aeternity/aepp-sdk/es/utils/aepp-wallet-communication/wallet-detector](api/utils/aepp-wallet-communication/wallet-detector.md)
   * [@aeternity/aepp-sdk/es/utils/crypto](api/utils/crypto.md)
   * [@aeternity/aepp-sdk/es/utils/keystore](api/utils/keystore.md)
   * [@aeternity/aepp-sdk/es/utils/swagger](api/utils/swagger.md)
