+++ docs/api/contract/aci.md
@@ -71,7 +71,7 @@ Generate contract ACI object with predefined js methods for contract usage - can
 | [options.aci] | `String` |  | Contract ACI |
 | [options.contractAddress] | `String` |  | Contract address |
 | [options.filesystem] | `Object` |  | Contact source external namespaces map |
-| [options.forceCodeCheck] | `Boolean` | <code>false</code> | Flag to force validation of corresponding on chain bytecode |
+| [options.forceCodeCheck] | `Object` |  | Don't check contract code |
 | [options.opt] | `Object` |  | Contract options |
 
 **Example**
