+++ es/utils/aepp-wallet-communication/helpers.js
@@ -1,3 +1,6 @@
+/**
+ * Browser helper functions
+ */
 /* eslint-disable no-undef */
 export const getBrowserAPI = () => {
   if (chrome && chrome.runtime) return chrome
@@ -10,6 +13,37 @@ export const isInIframe = () => window !== window.parent
 export const getWindow = () => {
   if (!window) throw new Error('Browser is not detected')
   return window
+}
+
+/**
+ * RPC helper functions
+ */
+export const sendMessage = (messageId, connection) => ({ id, method, params, result, error }, isNotificationOrResponse = false) => {
+  // Increment id for each request
+  isNotificationOrResponse || (messageId += 1)
+  id = isNotificationOrResponse ? (id || null) : messageId
+  const msgData = params
+    ? { params }
+    : result
+      ? { result }
+      : { error }
+  connection.sendMessage({
+    jsonrpc: '2.0',
+    ...id ? { id } : {},
+    method,
+    ...msgData
+  })
+  return id
+}
+
+export const receive = (handler, msgId) => (msg) => {
+  if (!msg || !msg.jsonrpc || msg.jsonrpc !== '2.0' || !msg.method) {
+    console.warn('Receive invalid message', msg)
+    return
+  }
+  // Increment id for each request
+  if (msg.id && +msg.id > msgId) msgId += 1
+  handler(msg)
 }
 
 export const getHandler = (schema, msg) => {
@@ -26,3 +60,5 @@ export const message = (method, params = {}) => ({ method, params })
 export const responseMessage = (id, method, { error, result } = {}) => ({ id, method, ...(error ? { error } : { result }) })
 
 export const sendResponseMessage = (client) => (id, method, data) => client.sendMessage(responseMessage(id, method, data), true)
+
+export const voidFn = () => undefined
