+++ es/utils/aepp-wallet-communication/rpc/wallet-rpc.js
@@ -1,13 +1,20 @@
+/**
+ * RPC handler for WAELLET side
+ *
+ * @module @aeternity/aepp-sdk/es/utils/aepp-wallet-communication/rpc/wallet-rpc
+ * @export WalletRpc
+ * @example import WalletRpc from '@aeternity/aepp-sdk/es/utils/aepp-wallet-communication/rpc/wallet-rpc'
+ */
 import Ae from '../../../ae'
 import Accounts from '../../../accounts'
 import Selector from '../../../account/selector'
 
-import { WalletClients } from './wallet-clients'
+import { RpcClients } from './rpc-clients'
 import { getBrowserAPI, getHandler, message, sendResponseMessage } from '../helpers'
 import { ERRORS, METHODS, RPC_STATUS, VERSION, WALLET_TYPE, SUBSCRIPTION_VALUES } from '../schema'
 import uuid from 'uuid/v4'
 
-const rpcClients = WalletClients()
+const rpcClients = RpcClients()
 
 const NOTIFICATIONS = {
   [METHODS.closeConnection]: (instance, { client }) =>
@@ -140,14 +147,33 @@ const handleMessage = (instance, id) => async (msg) => {
   }
 }
 
+/**
+ * Contain functionality for aepp interaction and managing multiple aepps
+ * @alias module:@aeternity/aepp-sdk/es/utils/aepp-wallet-communication/rpc/wallet-rpc
+ * @function
+ * @rtype Stamp
+ * @param {Object} param Init params object
+ * @param {String=} [param.name] Wallet name
+ * @param {Function} onConnection Call-back function for incoming AEPP connection (Second argument contain function for accept/deny request)
+ * @param {Function} onSubscription Call-back function for incoming AEPP account subscription (Second argument contain function for accept/deny request)
+ * @param {Function} onSign Call-back function for incoming AEPP sign request (Second argument contain function for accept/deny request)
+ * @param {Function} onAskAccounts Call-back function for incoming AEPP get address request (Second argument contain function for accept/deny request)
+ * @param {Function} onDisconnect Call-back function for disconnect event
+ * @return {Object}
+ */
 export const WalletRpc = Ae.compose(Accounts, Selector, {
-  init ({ icons, name, onConnection, onSubscription, onSign, onDisconnect, onAskAccounts }) {
+  init ({ name, onConnection, onSubscription, onSign, onDisconnect, onAskAccounts }) {
+    const eventsHandlers = ['onConnection', 'onSubscription', 'onSign', 'onDisconnect']
     // CallBacks for events
     this.onConnection = onConnection
     this.onSubscription = onSubscription
     this.onSign = onSign
     this.onDisconnect = onDisconnect
     this.onAskAccounts = onAskAccounts
+
+    eventsHandlers.forEach(event => {
+      if (typeof this[event] !== 'function') throw new Error(`Call-back for ${event} must be an function!`)
+    })
     //
     this.name = name
     this.id = uuid()
@@ -191,9 +217,24 @@ export const WalletRpc = Ae.compose(Accounts, Selector, {
     }
   },
   methods: {
+    /**
+     * Get RpcClients object which contain all connected AEPPS
+     * @function getClients
+     * @instance
+     * @rtype () => Object
+     * @return {Object}
+     */
     getClients () {
       return rpcClients
     },
+    /**
+     * Add new AEPP connection
+     * @function addRpcClient
+     * @instance
+     * @rtype (clientConnection: Object) => Object
+     * @param {Object} clientConnection AEPP connection object
+     * @return {void}
+     */
     addRpcClient (clientConnection) {
       // @TODO  detect if aepp has some history based on origin????: if yes use this instance for connection
       const id = uuid()
@@ -207,12 +248,28 @@ export const WalletRpc = Ae.compose(Accounts, Selector, {
         }
       )
     },
+    /**
+     * Share wallet info
+     * Send shareWalletInfo message to notify AEPP about wallet
+     * @function shareWalletInfo
+     * @instance
+     * @rtype (postFn: Function) => void
+     * @param {Function} postFn Send message function like `(msg) => void`
+     * @return {void}
+     */
     shareWalletInfo (postFn) {
       postFn({
         jsonrpc: '2.0',
         ...message(METHODS.wallet.readyToConnect, { ...this.getWalletInfo() })
       })
     },
+    /**
+     * Get Wallet info object
+     * @function getWalletInfo
+     * @instance
+     * @rtype () => Object
+     * @return {Object} Object with wallet information(id, name, network, ...)
+     */
     getWalletInfo () {
       return {
         id: getBrowserAPI().runtime.id || this.id,
@@ -222,6 +279,13 @@ export const WalletRpc = Ae.compose(Accounts, Selector, {
         type: getBrowserAPI().runtime.id ? WALLET_TYPE.extension : WALLET_TYPE.window
       }
     },
+    /**
+     * Get Wallet accounts
+     * @function getAccounts
+     * @instance
+     * @rtype () => Object
+     * @return {Object} Object with accounts information({ connected: Object, current: Object })
+     */
     getAccounts () {
       return {
         current: this.Selector.address ? { [this.Selector.address]: {} } : {},
