+++ examples/browser/extension/package.json
@@ -8,7 +8,7 @@
   },
   "devDependencies": {
     "clean-webpack-plugin": "^1.0.0",
-    "copy-webpack-plugin": "^4.6.0",
+    "copy-webpack-plugin": "^5.1.1",
     "css-loader": "2.1.0",
     "file-loader": "3.0.1",
     "fs-extra": "7.0.1",
@@ -16,11 +16,11 @@
     "html-webpack-plugin": "3.2.0",
     "style-loader": "^0.23.1",
     "webpack": "4.28.4",
+    "webpack-cli": "^3.2.1",
     "webpack-dev-server": "3.1.14",
-    "write-file-webpack-plugin": "4.5.0",
-    "webpack-cli": "^3.2.1"
+    "write-file-webpack-plugin": "4.5.0"
   },
   "dependencies": {
-    "@aeternity/aepp-sdk": "https://github.com/aeternity/aepp-sdk-js#feature/aex-2"
+    "@aeternity/aepp-sdk": "https://github.com/aeternity/aepp-sdk-js#feat/switch-to-post-message-for-extension"
   }
 }
