+++ examples/browser/extension/src/js/background.js
@@ -4,15 +4,15 @@ import '../img/icon-34.png'
 import MemoryAccount from '@aeternity/aepp-sdk/es/account/memory'
 import { RpcWallet } from '@aeternity/aepp-sdk/es/ae/wallet'
 import BrowserRuntimeConnection
-  from '@aeternity/aepp-sdk/es/utils/aepp-wallet-communication/wallet-connection/browser-runtime'
-import { generateKeyPair } from '../../../../../es/utils/crypto'
+  from '@aeternity/aepp-sdk/es/utils/aepp-wallet-communication/connection/browser-runtime'
+import { generateKeyPair } from '@aeternity/aepp-sdk/es/utils/crypto'
 
-const account = MemoryAccount({
-  keypair: {
-    secretKey: 'YOUR_PRIV',
-    publicKey: 'YOUR_PUB'
-  }
-})
+// const account = MemoryAccount({
+//   keypair: {
+//     secretKey: 'YOUR_PRIV',
+//     publicKey: 'YOUR_PUB'
+//   }
+// })
 
 const account2 = MemoryAccount({ keypair: generateKeyPair() })
 
@@ -48,17 +48,9 @@ const accounts = [
   //         }
   //     }
   // })(),
-  account,
+  // account,
   account2
 ]
-//
-const postToContent = (data) => {
-  chrome.tabs.query({}, function (tabs) { // TODO think about direct communication with tab
-    const message = { method: 'pageMessage', data }
-    tabs.forEach(({ id }) => chrome.tabs.sendMessage(id, message)) // Send message to all tabs
-  })
-}
-
 // Send wallet connection info to Aepp throug content script
 const NODE_URL = 'https://sdk-testnet.aepps.com'
 const NODE_INTERNAL_URL = 'https://sdk-testnet.aepps.com'
@@ -105,17 +97,15 @@ RpcWallet({
     }
   }
 }).then(wallet => {
-  // Subscribe for runtime connection
-  chrome.runtime.onConnectExternal.addListener(async (port) => {
+  chrome.runtime.onConnect.addListener(async function (port) {
     // create Connection
     const connection = await BrowserRuntimeConnection({ connectionInfo: { id: port.sender.frameId }, port })
     // add new aepp to wallet
     wallet.addRpcClient(connection)
+    // Share wallet details
+    wallet.shareWalletInfo(port.postMessage.bind(port))
+    setTimeout(() => wallet.shareWalletInfo(port.postMessage.bind(port)), 3000)
   })
-  // Share wallet info with extensionId to the page
-  debugger
-  // Send wallet connection info to Aepp throug content script
-  setInterval(() => wallet.shareWalletInfo(postToContent), 5000)
 }).catch(err => {
   console.error(err)
 })
